refactor(web-ui): split shared api contracts
